+++ src/onegov.pdf/onegov/pdf/__init__.py
@@ -1,8 +1,14 @@
+from onegov.pdf.pdf import page_fn_footer
+from onegov.pdf.pdf import page_fn_header
+from onegov.pdf.pdf import page_fn_header_and_footer
 from onegov.pdf.pdf import Pdf
 from onegov.pdf.signature import LexworkSigner
 
 
 __all__ = [
+    'LexworkSigner',
+    'page_fn_footer',
+    'page_fn_header_and_footer',
+    'page_fn_header',
     'Pdf',
-    'LexworkSigner'
 ]